requirements.txt
alembic==1.13.1
annotated-types==0.7.0
appnope==0.1.4
asttokens==2.4.1
cachetools==5.4.0
chardet==5.2.0
cloudpickle==3.0.0
cmaes==0.10.0
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
executing==2.0.1
filelock==3.15.4
fonttools==4.51.0
greenlet==3.0.3
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.3
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
llvmlite==0.42.0
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numba==0.59.1
numpy==1.26.4
optuna==3.6.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pyproject-api==1.7.1
pytest==8.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
shap==0.45.1
six==1.16.0
slicer==0.0.8
SQLAlchemy==2.0.31
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4
tox==4.16.0
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.3
wcwidth==0.2.13
widgetsnbextension==4.0.11
xgboost==2.0.3
sagemaker
